yage-gl: move functions to "functions" module
